Convert I18n Module to TS
