--- src/i18n/config.js
@@ -1,3 +1,0 @@
-let EVENTS = ["i18n:add", "i18n:update", "i18n:locale"];
-
-module.exports = { EVENTS };
