+++ src/i18n/config.ts
@@ -1,0 +1,1 @@
+export const EVENTS = ["i18n:add", "i18n:update", "i18n:locale"];
